Add TX and RX timeout options.
This allows the program to be run for a fixed time and then exit. The exit code
will be the difference between the number of sent and received bytes, plus the
number of errors, i.e. zero if everything was correct, meaning that this can be
used as part of an automated test script.