Improve timeout reporting.
Previously a message would be printed every second if no data was being
received. This commit improves things so that it prints messages for
both transmission and reception timing out, and tells you how long it
has been since data was last received/transmitted. The initial timeout
for both transmission and reception is now two seconds, and after that
messages are printed every second.