Add ASCII byte range option.
This adds a new option to have the output bytes range from 32 to 126
instead of the default 0 to 255. There are two reasons why this is
useful, one is that you can directly output the bytes at the receiving
end without getting any unprintable codes, and the other is that the
sequence is not a power of two in length, which would reveal problems
where a faulty serial driver dropped a complete buffer of data (e.g. 256
bytes or 1024 bytes) that would get missed using the default sequence.