propagate error code and exit with negative errno
On failure, "1" was returned by the program. This is problematic when
checking for proper execution, since this value can also be returned by
the count at the end of the program.

Instead, we force all return code to be negative. In addition, we
propagate the error code when it was available from another command
(ioctl mainly). When it was not availabe, i.e. it is an error condition
within the program checks, we exit with an errno which seemed
appropriate for the condition.

In the console, those return values are casted from int to unsigned int,
therefore, a return code of `-1` will give a return code of 255 when
using `$?` in a command line. The program uses return code 1 to 125 to
indicate that there was a transmission error detected (either an error
on the value received, or a difference in the number of characters
received vs transmitted). Therefore, to check in command line if the
program exitted abnormally, the user should check against `rc > 125`.

Hopefully, this allows better analysis of the program's termination
state and location of the source of test failures.

Signed-off-by: Laurier Loiselle <lal@xiphos.com>