compute error count valid when no_tx or no_rx
When no tx or no rx parameters are set, the returned count will saturate
because the tx or rx value will be 0. This makes the returned value
useless to check directly if there was a transmission error.

In that case, what needs to be checked is the error count, which informs
the application on reception error.

When doing only tx, the return should always be 0, since there is no
reception error in that case. We didn't add a specific case for this
to keep the change less verbose.

Signed-off-by: Laurier Loiselle <lal@xiphos.com>