add option to flush I/O buffers
Add `--flush-buffers` command line option to allow to flush buffers before
starting the main loop.  It flushes both the data received but not read and the
data written.

This allows to get rid of some desynchronization errors that could happen when
the sender sends data before the receiver is ready.

A delay of 100ms has been introduced before sending TCIOFLUSH ioctl command in
order to let data arrive in the buffer before flushing.

Signed-off-by: Sandra Tobajas <sto@xiphos.com>