add rx threashold to avoid infinite loop condition
Previously it was possible to be stuck in an infinite loop when using full duplex transmission.
The read side previously required 1024 bytes before it would continue. This can lead to a
situation where both devices are stuck in the read loop as more bytes are required to meet the
exit condition. The tx side does not always send 1024 bytes, particularly if the tx-bytes option
is used. This can also prevent uni-directional transfer tests from completing if the transmitter
stops before the receiver.

The rx_bytes_threash, allows for the condition to be adjusted to avoid entering the infinite
loop condition.

Signed-off-by: Andrew Cramp <acr@xiphos.com>