linux-serial-test.c: fix returned error code
_cl_no_rx/_cl_no_rx are true in two cases, first when the relevant
command line paramter is set and second when a rx or tx time is set
and that time expired.

This fixes the second case in a loopback test. With the change the
application now correctly returns 125 if the number of written and
read chars differ.

E.g. `./linux-serial-test -o2 -i3 -b 115200 -p /dev/ttyS1 ; echo $?`
returns now 0 with a loopback and 125 without, before it returned 0
with or without a loopback.

Fixes: 4e57f58c58ae ("compute error count valid when no_tx or no_rx")
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>