linux-serial-test.c: fix potential hang in while loop
process_read_data() assumes that we can always wait for reception
of 1024 chars. However that is not true if one sets the number of
chars with the '-w' cmdline parameter or chars are lost. Maybe there
are other reasons.

Replace the magic number of 1024 by calculating the number of expected
chars from _cl_tx_bytes.

Brake a possible infinite while loop by adding a timeout to the loop
calculated from the expected chars times chartime.

Fixes: 7fd1057f8a95 ("Add loop to read all data in rcv buffer")
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>