Remove useless(?) while loop
A while loop was added by 7fd1057f8a95db4e79393c0ba78fccee0752afa5
that caused the program to hang. The reason for the loop is not clear.
The commit comment had said "Previously, read would be called only
once", but that isn't the case now as process_read_data() is called in
a while loop from main().