Some serial drivers crash with valid BOTHER baudrate requests
speed_t is unsigned int, so it should allow requesting speeds up to
UINT_MAX (~4e9). However, the ftdi_sio driver currently crashes
unrepairably when speeds over INT_MAX (~2e9) are requsted.

Unrepairably meaning the driver doesn't just hang causing the process
to get stuck in a device wait, but there is no way to reset it. In
fact, even powering off or rebooting the machine fails for me and I
have to hold down the power button for 30 seconds.