Do not return an error if _read != _write_count
#ifdef'd out an unnecessarily pedantic and incorrect check.

The _write_count can be inflated due to buffering in the driver. It is
not correct to exit with an error simply because more data was written
than read until we figure out how to unbuffer the writes. (Or, is it
possible to measure the bytes in the buffer?)

Fortunately, being pedantic about the read and write counts matching
exactly is not helpful for testing a serial device. Any case which it
would catch is already caught better by clearer and simpler tests:

  * If bytes at the beginning or middle of the transmission are
  missing, then process_read_data() will catch it due to the data
  mismatch.

  * If no bytes are received, an error is returned with a suggestion
  to check that the loopback is plugged in.

  * If the above errors haven't triggered, we know that we have
  received *some* data and it was valid. A hypothetical device that
  fails after receiving some number n of bytes would be detected as
  having a lower estimated baudrate and, if it exceeded the tolerance
  threshold, an error would be returned.

  (Note that when n is close to the number of bytes sent, tx_count, an
  error would _not_ be detected. However, it is only a slight increase
  to the always present chance that an error would be missed due to n
  being greater than or equal to tx_count.)